requirements.txt
astroid==2.1.0
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
coinmarketcap==5.0.3
colorama==0.4.1
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.1
django-graphos==0.3.41
django-heroku==0.3.1
djongo==1.2.32
gunicorn==19.9.0
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
numpy==1.13.3
pandas==0.23.4
pipenv==2018.11.26
praw==5.2.0
prawcore==0.12.0
psycopg2==2.8.2
public==2018.11.20
PyJWT==1.5.3
pylint==2.2.2
pymongo==3.4.0
PySocks==1.6.7
python-dateutil==2.7.5
pytz==2017.3
requests==2.20.1
requests-cache==0.4.13
six==1.11.0
sqlparse==0.2.4
timedelta==2018.11.20
twilio==6.9.0
typed-ast==1.3.1
update-checker==0.16
urllib3==1.22
virtualenv==16.5.0
virtualenv-clone==0.5.3
whitenoise==4.1.2
wrapt==1.11.1
